chore: move KVPbApi impl to seperate files
