+++ src/meta/api/src/kv_pb_api/errors/mod.rs
@@ -1,0 +1,26 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Defines errors used by protobuf based API.
+
+mod decode_error;
+mod encode_error;
+mod read_error;
+mod write_error;
+
+pub use self::decode_error::PbDecodeError;
+pub use self::encode_error::PbEncodeError;
+pub use self::read_error::NoneValue;
+pub use self::read_error::PbApiReadError;
+pub use self::write_error::PbApiWriteError;
