+++ src/meta/api/src/kv_pb_api/mod.rs
@@ -14,8 +14,9 @@
 
 //! Kv API with `kvapi::Key` type key and protobuf encoded value.
 
+pub mod errors;
+
 mod codec;
-mod errors;
 mod upsert_pb;
 
 use std::future::Future;
@@ -25,115 +26,23 @@ use databend_common_meta_kvapi::kvapi::DirName;
 use databend_common_meta_kvapi::kvapi::KVApi;
 use databend_common_meta_kvapi::kvapi::Key;
 use databend_common_meta_kvapi::kvapi::NonEmptyItem;
-use databend_common_meta_types::protobuf::StreamItem;
 use databend_common_meta_types::Change;
-use databend_common_meta_types::InvalidReply;
-use databend_common_meta_types::MetaError;
-use databend_common_meta_types::MetaNetworkError;
 use databend_common_meta_types::SeqV;
 use databend_common_meta_types::UpsertKV;
 use databend_common_proto_conv::FromToProto;
-use databend_common_proto_conv::Incompatible;
 use futures::future::FutureExt;
 use futures::future::TryFutureExt;
 use futures::stream::BoxStream;
 use futures::stream::StreamExt;
 use futures::TryStreamExt;
-use PbApiReadError::KvApiError;
 
+pub(crate) use self::codec::decode_non_empty_item;
 pub(crate) use self::codec::decode_seqv;
 pub(crate) use self::codec::decode_transition;
 pub(crate) use self::codec::encode_operation;
-pub use self::errors::PbApiWriteError;
-pub use self::errors::PbEncodeError;
 pub use self::upsert_pb::UpsertPB;
-
-// TODO: move error to separate file
-
-/// An error occurred when decoding protobuf encoded value.
-#[derive(Clone, Debug, PartialEq, thiserror::Error)]
-#[error("PbDecodeError: {0}")]
-pub enum PbDecodeError {
-    DecodeError(#[from] prost::DecodeError),
-    Incompatible(#[from] Incompatible),
-}
-
-impl From<PbDecodeError> for MetaError {
-    fn from(value: PbDecodeError) -> Self {
-        match value {
-            PbDecodeError::DecodeError(e) => MetaError::from(InvalidReply::new("", &e)),
-            PbDecodeError::Incompatible(e) => MetaError::from(InvalidReply::new("", &e)),
-        }
-    }
-}
-
-/// An error occurs when found an unexpected None value.
-#[derive(Clone, Debug, PartialEq, thiserror::Error)]
-#[error("NoneValue: unexpected None value of key: '{key}'")]
-pub struct NoneValue {
-    key: String,
-}
-
-impl NoneValue {
-    pub fn new(key: impl ToString) -> Self {
-        NoneValue {
-            key: key.to_string(),
-        }
-    }
-}
-
-/// An error occurs when reading protobuf encoded value from kv store.
-#[derive(Clone, Debug, PartialEq, thiserror::Error)]
-#[error("PbApiReadError: {0}")]
-pub enum PbApiReadError<E> {
-    DecodeError(#[from] prost::DecodeError),
-    Incompatible(#[from] Incompatible),
-    KeyError(#[from] kvapi::KeyError),
-    NoneValue(#[from] NoneValue),
-    /// Error returned from KVApi.
-    KvApiError(E),
-}
-
-impl<E> From<PbDecodeError> for PbApiReadError<E> {
-    fn from(e: PbDecodeError) -> Self {
-        match e {
-            PbDecodeError::DecodeError(e) => PbApiReadError::DecodeError(e),
-            PbDecodeError::Incompatible(e) => PbApiReadError::Incompatible(e),
-        }
-    }
-}
-
-impl From<PbApiReadError<MetaError>> for MetaError {
-    /// For KVApi that returns MetaError, convert protobuf related error to MetaError directly.
-    ///
-    /// Because MetaError contains network protocol level error variant.
-    /// If there is a decoding error, consider it as network level error.
-    fn from(value: PbApiReadError<MetaError>) -> Self {
-        match value {
-            PbApiReadError::DecodeError(e) => {
-                let inv = InvalidReply::new("", &e);
-                let net_err = MetaNetworkError::InvalidReply(inv);
-                MetaError::NetworkError(net_err)
-            }
-            PbApiReadError::Incompatible(e) => {
-                let inv = InvalidReply::new("", &e);
-                let net_err = MetaNetworkError::InvalidReply(inv);
-                MetaError::NetworkError(net_err)
-            }
-            PbApiReadError::KeyError(e) => {
-                let inv = InvalidReply::new("", &e);
-                let net_err = MetaNetworkError::InvalidReply(inv);
-                MetaError::NetworkError(net_err)
-            }
-            PbApiReadError::NoneValue(e) => {
-                let inv = InvalidReply::new("", &e);
-                let net_err = MetaNetworkError::InvalidReply(inv);
-                MetaError::NetworkError(net_err)
-            }
-            KvApiError(e) => e,
-        }
-    }
-}
+use crate::kv_pb_api::errors::PbApiReadError;
+use crate::kv_pb_api::errors::PbApiWriteError;
 
 /// This trait provides a way to access a kv store with `kvapi::Key` type key and protobuf encoded value.
 pub trait KVPbApi: KVApi {
@@ -209,7 +118,10 @@ pub trait KVPbApi: KVApi {
     {
         let key = key.to_string_key();
         async move {
-            let raw_seqv = self.get_kv(&key).await.map_err(KvApiError)?;
+            let raw_seqv = self
+                .get_kv(&key)
+                .await
+                .map_err(PbApiReadError::KvApiError)?;
             let v = raw_seqv.map(decode_seqv::<K::ValueType>).transpose()?;
             Ok(v)
         }
@@ -269,7 +181,10 @@ pub trait KVPbApi: KVApi {
     {
         let prefix = prefix.to_string_key();
         async move {
-            let strm = self.list_kv(&prefix).await.map_err(KvApiError)?;
+            let strm = self
+                .list_kv(&prefix)
+                .await
+                .map_err(PbApiReadError::KvApiError)?;
             let strm = strm.map(decode_non_empty_item::<K, Self::Error>);
             Ok(strm.boxed())
         }
@@ -277,40 +192,3 @@ pub trait KVPbApi: KVApi {
 }
 
 impl<T> KVPbApi for T where T: KVApi + ?Sized {}
-
-/// Decode key and protobuf encoded value from `StreamItem`.
-///
-/// It requires K to be static because it is used in a static stream map()
-fn decode_non_empty_item<K, E>(
-    r: Result<StreamItem, E>,
-) -> Result<NonEmptyItem<K>, PbApiReadError<E>>
-where
-    K: kvapi::Key + 'static,
-    K::ValueType: FromToProto,
-{
-    match r {
-        Ok(item) => {
-            let k = K::from_str_key(&item.key)?;
-
-            let raw = item.value.ok_or_else(|| NoneValue::new(item.key))?;
-            let v = decode_seqv::<K::ValueType>(SeqV::from(raw))?;
-
-            Ok(NonEmptyItem::new(k, v))
-        }
-        Err(e) => Err(KvApiError(e)),
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::kv_pb_api::PbDecodeError;
-
-    #[test]
-    fn test_error_message() {
-        let e = PbDecodeError::DecodeError(prost::DecodeError::new("decode error"));
-        assert_eq!(
-            "PbDecodeError: failed to decode Protobuf message: decode error",
-            e.to_string()
-        );
-    }
-}
